requirements.txt
streamlit>=1.28.0
groq
python-dotenv
